# improve batch_norm contiguous case's performance