--- a/aten/src/ATen/native/Normalization.cpp
+++ b/aten/src/ATen/native/Normalization.cpp
@@ -8,12 +8,15 @@
 #include <ATen/detail/CUDAHooksInterface.h>
 #include <ATen/native/TensorIterator.h>
 #include <ATen/native/cpu/Loops.h>
+#include <ATen/native/batch_norm.h>
 
 #include <vector>
 
 static const int MIOPEN_DIM_MAX = 5;
 
 namespace at { namespace native {
+
+DEFINE_DISPATCH(batch_norm_cpu_inference_contiguous_stub);
 
 namespace {
   void check_dims_match_num_input_features(const char* arg_name, int64_t expected, int64_t actual){
@@ -87,59 +90,6 @@ void batch_norm_cpu_inference_collect_linear_and_constant_terms(
   }
 }
 
-/// A fast path for CPU inference when all tensors are contiguous.
-/// This code achieves machine bandwidth peak without AVX support.
-/// If this changes for future architectures, we can move it to the cpu/
-/// directory.
-template<typename scalar_t>
-void batch_norm_cpu_inference_contiguous(Tensor& output, const Tensor& input,
-    const Tensor& weight /* optional */, const Tensor& bias /* optional */,
-    const Tensor& mean, const Tensor& variance, double eps) {
-
-  int64_t n_batch = input.size(0);
-  int64_t n_channel = input.size(1);
-  int64_t image_size = input.numel() / n_batch / n_channel;
-
-  scalar_t* output_data = output.data_ptr<scalar_t>();
-  const scalar_t* input_data = input.data_ptr<scalar_t>();
-
-  Tensor alpha = at::empty_like(mean, LEGACY_CONTIGUOUS_MEMORY_FORMAT);
-  Tensor beta = at::empty_like(mean, LEGACY_CONTIGUOUS_MEMORY_FORMAT);
-  scalar_t* alpha_data = alpha.data_ptr<scalar_t>();
-  scalar_t* beta_data = beta.data_ptr<scalar_t>();
-
-  batch_norm_cpu_inference_collect_linear_and_constant_terms<scalar_t>(
-      alpha_data, beta_data, n_channel, weight, bias, mean, variance, eps);
-
-  // Apply the linear terms to the input,
-  // output(n, c, h, w) = input(n, c, h, w) * alpha(c) + beta(c)
-  // No need to use parallel_for as this function is supposed to be
-  // memory-limited.
-  // Keep the loop struture simple to make sure compiler vectorization kicks in.
-  if (image_size != 1) {
-    for (int64_t n = 0; n < n_batch; ++n) {
-      for (int64_t c = 0; c < n_channel; ++c) {
-        for (int64_t i = 0; i < image_size; ++i) {
-          // Keep all the offset calculation within the inner loop for
-          // simplicity. Compilers are very good at hoisting the common part
-          // outside.
-          int64_t offset = n * n_channel * image_size + c * image_size + i;
-          output_data[offset] = input_data[offset] * alpha_data[c] +
-              beta_data[c];
-        }
-      }
-    }
-  } else {
-    // image_size == 1
-    for (int64_t n = 0; n < n_batch; ++n) {
-      for (int64_t c = 0; c < n_channel; ++c) {
-        int64_t offset = n * n_channel + c;
-        output_data[offset] = input_data[offset] * alpha_data[c] + beta_data[c];
-      }
-    }
-  }
-}
-
 /// A fast path for CPU inference when all tensors are channels last contiguous.
 /// This code achieves machine bandwidth peak without AVX support.
 /// If this changes for future architectures, we can move it to the cpu/
@@ -207,8 +157,8 @@ std::tuple<Tensor,Tensor,Tensor> batch_norm_cpu_transform_input_template(
       && running_var.is_contiguous()) {
 
     Tensor output = at::empty_like(input, LEGACY_CONTIGUOUS_MEMORY_FORMAT);
-    batch_norm_cpu_inference_contiguous<scalar_t>(
-      output, input, weight, bias, running_mean, running_var, eps);
+    batch_norm_cpu_inference_contiguous_stub(kCPU, output, input, weight,
+        bias, running_mean, running_var, eps);
     return std::make_tuple(output, save_mean, save_invstd);
   }